separate gui code into  separate file
